# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Aeko Enhanced API",
    "version": "2.0.0",
...
    "contact": {
      "name": "Aeko Development Team",
      "email": "dev@aeko.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/interests": {
      "post": {
        "summary": "Create a new interest (Admin only)",
        "tags": [
          "Interests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "displayName"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "displayName": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "icon": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Interest created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Interest already exists"
          }
        }
      },
      "get": {
        "summary": "Get all active interests",
        "tags": [
          "Interests"
        ],
        "responses": {
          "200": {
            "description": "List of active interests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Interest"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/interests/{id}": {
      "put": {
        "summary": "Update an interest (Admin only)",
        "tags": [
          "Interests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "displayName": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "icon": {
                    "type": "string"
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Interest updated successfully"
          },
          "404": {
            "description": "Interest not found"
          }
        }
      },
      "delete": {
        "summary": "Delete an interest (Admin only)",
        "tags": [
          "Interests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Interest deleted successfully"
          },
          "404": {
            "description": "Interest not found"
          }
        }
      }
    },
...
    "/api/user/interests": {
      "get": {
        "summary": "Get user's interests",
        "tags": [
          "User Interests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User's interests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Interest"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Update user's interests",
        "tags": [
          "User Interests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "interestIds"
                ],
                "properties": {
                  "interestIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Array of interest IDs"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Interests updated successfully"
          },
          "400": {
            "description": "Invalid interest IDs"
          }
        }
      }
    },
    "/api/user/interests/{interestId}": {
      "delete": {
        "summary": "Remove an interest from user's interests",
        "tags": [
          "User Interests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "interestId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Interest removed successfully"
          },
          "404": {
            "description": "Interest not found in user's interests"
          }
        }
      }
    },
...
  }
}